+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Removed
+
+- **[BREAKING]** Westpac QuickWeb payment event
+
 ## [12.0.1] - 2024-06-04
 
 ### Changed